updated app router, removed extra app
