+++ apps/frontend/src/app/layout.tsx
@@ -5,10 +5,11 @@ import type { Viewport } from "next"
 import { GoogleOAuthProvider } from "@react-oauth/google"
 import { Inter, Source_Serif_4, JetBrains_Mono } from "next/font/google"
 
-// import localFont from "next/font/local"
-import classNames from "@/src/utils/classNames"
-
 import "../styles/main.css"
+import Sidebar from "@/src/components/Sidebar"
+import { AuthProvider } from "@/src/contexts/AuthContext"
+import QueryProvider from "@/src/contexts/QueryProvider"
+import classNames from "@/src/utils/classNames"
 
 const sansFont = Inter({
   variable: "--sans-font",
@@ -28,23 +29,25 @@ const monoFont = JetBrains_Mono({
   weight: ["100", "200", "300", "400", "500", "600", "700", "800"],
 })
 
-// const monoFont = localFont({
-//   variable: "--mono-font",
-//   src: [
-//     {
-//       path: "@/public/fonts/JetBrainsMono-Regular.ttf",
-//       weight: "regular",
-//       style: "normal",
-//     },
-//   ],
-// })
-
 export const viewport: Viewport = {
   themeColor: "#000000",
 }
 
 interface Props {
   children: React.ReactNode
+}
+
+const AppLayout: React.FC<Props> = ({ children }) => {
+  return (
+    <AuthProvider>
+      <QueryProvider>
+        <main className="flex h-screen bg-background">
+          <Sidebar />
+          <section className="flex-1">{children}</section>
+        </main>
+      </QueryProvider>
+    </AuthProvider>
+  )
 }
 
 const RootLayout: React.FC<Props> = ({ children }) => {
@@ -61,7 +64,7 @@ const RootLayout: React.FC<Props> = ({ children }) => {
         <GoogleOAuthProvider
           clientId={process.env.NEXT_PUBLIC_GOOGLE_CLIENT_ID ?? ""}
         >
-          {children}
+          <AppLayout>{children}</AppLayout>
         </GoogleOAuthProvider>
       </body>
     </html>
