+++ apps/frontend/src/app/settings/[slug]/page.tsx
