+++ apps/frontend/src/app/space/[slug]/page.tsx
