+++ apps/frontend/src/app/space/notes/page.tsx
