+++ apps/frontend/src/app/space/page.tsx
